Refactored interfaces imports
